GDA feature set expanded - 82% accuracy
the feature extractor was expanded to compute not only intensity mean and std dev for each frame, but also for each of R, G and B, extending x into R^8. GDA was also modified to handle a feature vector in any dimension. accuracy using rekall labels for training and true labels for testing reached 82%.